requirements.txt
aiohttp
trafilatura
nest_asyncio
langgraph
openai
google-generativeai
python-dotenv
fastapi
uvicorn
pytesseract
Pillow
python-multipart